Better strictfp test
